correction constant right

+++ include/ffb.h
@@ -27,7 +27,7 @@ typedef struct
     uint8_t startByte;            // Start byte
     double spring;                // from 0x00 to 0x7F -> 128 levels converted to double 0.0 -> 1.0
     double friction;              // from 0x00 to 0x7F -> 128 levels converted to double 0.0 -> 1.0 
-    uint8_t torqueDirection;      // 0x00 = Left, 0x01  = Right
+    int torqueDirection;          // 0x00 = Left, 0x01  = Right
     double torquePower;           // from 0x00 to 0xFF -> 128 levels converted to double 0.0 -> 1.0
     uint8_t crc;                  // (D1 ^ D2 ^ D3 ^ D4) & 0x7F
 } FFBPacket;
